+++ resources/leiningen/new/luminus/db/src/datomic.clj
@@ -1,0 +1,48 @@
+(ns <<project-ns>>.db.core
+  (:require [datomic.api :as d]
+            [mount.core :refer [defstate]]
+            [<<project-ns>>.config :refer [env]]))
+
+(defstate conn
+          :start (-> env :database-url d/connect)
+          :stop (-> conn .release))
+
+(defn create-schema []
+  (let [schema [{:db/ident              :user/id
+                 :db/valueType          :db.type/string
+                 :db/cardinality        :db.cardinality/one
+                 :db.install/_attribute :db.part/db}
+                {:db/ident              :user/first-name
+                 :db/valueType          :db.type/string
+                 :db/cardinality        :db.cardinality/one
+                 :db.install/_attribute :db.part/db}
+                {:db/ident              :user/last-name
+                 :db/valueType          :db.type/string
+                 :db/cardinality        :db.cardinality/one
+                 :db.install/_attribute :db.part/db}
+                {:db/ident              :user/email
+                 :db/valueType          :db.type/string
+                 :db/cardinality        :db.cardinality/one
+                 :db.install/_attribute :db.part/db}]]
+    @(d/transact conn schema)))
+
+(defn entity [conn id]
+  (d/entity (d/db conn) id))
+
+(defn touch [conn results]
+  "takes 'entity ids' results from a query
+    e.g. '#{[272678883689461] [272678883689462] [272678883689459] [272678883689457]}'"
+  (let [e (partial entity conn)]
+    (map #(-> % first e d/touch) results)))
+
+(defn add-user [conn {:keys [id first-name last-name email]}]
+  @(d/transact conn [{:db/id           id
+                      :user/first-name first-name
+                      :user/last-name  last-name
+                      :user/email      email}]))
+
+(defn find-user [conn id]
+  (let [orders (d/q '[:find ?e :in $ ?id
+                      :where [?e :user/id ?id]]
+                    (d/db conn) id)]
+    (touch conn user)))